+++ crux-llvm/src/Crux/LLVM/Simulate.hs
@@ -59,7 +59,7 @@ import Lang.Crucible.Simulator.Profiling ( Metric(Metric) )
 
 
 -- crucible-llvm
-import Lang.Crucible.LLVM(llvmExtensionImpl, llvmGlobals, registerModuleFn )
+import Lang.Crucible.LLVM(llvmExtensionImpl, llvmGlobals, registerLazyModule )
 import Lang.Crucible.LLVM.Bytes ( bytesToBV )
 import Lang.Crucible.LLVM.Globals
         ( initializeAllMemory, populateAllGlobals )
@@ -74,7 +74,7 @@ import Lang.Crucible.LLVM.MemModel.CallStack (ppCallStack)
 import Lang.Crucible.LLVM.MemType (MemType(..), SymType(..), i8, memTypeAlign, memTypeSize)
 import Lang.Crucible.LLVM.Translation
         ( translateModule, ModuleTranslation, globalInitMap
-        , transContext, cfgMap, llvmPtrWidth, llvmTypeCtx
+        , transContext, getTranslatedCFG, llvmPtrWidth, llvmTypeCtx
         , LLVMTranslationWarning(..)
         )
 import Lang.Crucible.LLVM.Intrinsics
@@ -125,6 +125,8 @@ parseLLVM file =
        Right m  -> return m
 
 registerFunctions ::
+  Crux.Logs msgs =>
+  Log.SupportsCruxLLVMLogMessage msgs =>
   (ArchOk arch, IsSymInterface sym, HasLLVMAnn sym, ptrW ~ ArchWidth arch) =>
   LLVMOptions ->
   LLVM.Module ->
@@ -147,7 +149,7 @@ registerFunctions llvmOpts llvm_module mtrans fs0 =
        llvm_ctx
 
      -- register all the functions defined in the LLVM module
-     mapM_ (registerModuleFn llvm_ctx) $ Map.elems $ cfgMap mtrans
+     registerLazyModule sayTranslationWarning mtrans
 
 simulateLLVMFile ::
   Crux.Logs msgs =>
@@ -241,7 +243,7 @@ prepLLVMModule :: IsSymBackend sym bak
                -> IO (PreppedLLVM sym)
 prepLLVMModule llvmOpts halloc bak bcFile memVar = do
     llvmMod <- parseLLVM bcFile
-    (Some trans, warns) <-
+    Some trans <-
         let ?transOpts = transOpts llvmOpts
          in translateModule halloc memVar llvmMod
     mem <- let llvmCtxt = trans ^. transContext in
@@ -250,9 +252,8 @@ prepLLVMModule llvmOpts halloc bak bcFile memVar = do
              in llvmPtrWidth llvmCtxt $ \ptrW ->
                withPtrWidth ptrW $ do
                Log.sayCruxLLVM (Log.UsingPointerWidthForFile (intValue ptrW) (Text.pack bcFile))
-               populateAllGlobals bak (globalInitMap trans)
+               populateAllGlobals bak (trans ^. globalInitMap)
                  =<< initializeAllMemory bak llvmCtxt llvmMod
-    mapM_ sayTranslationWarning warns
     return $ PreppedLLVM llvmMod (Some trans) memVar mem
 
 sayTranslationWarning ::
@@ -276,21 +277,22 @@ checkFun ::
   GlobalVar Mem ->
   OverM personality sym LLVM ()
 checkFun llvmOpts trans memVar =
-  case Map.lookup (fromString nm) mp of
-    Just (_, AnyCFG anyCfg) ->
-      case cfgArgTypes anyCfg of
-        Empty -> simulateFun anyCfg emptyRegMap
+  liftIO (getTranslatedCFG trans (fromString nm)) >>= \case
+    Just (_, AnyCFG anyCfg, warns) ->
+      do liftIO (mapM_ sayTranslationWarning warns)
+         case cfgArgTypes anyCfg of
+           Empty -> simulateFun anyCfg emptyRegMap
 
-        (Empty :> LLVMPointerRepr w :> PtrRepr)
-          |  isMain, shouldSupplyMainArguments
-          ,  Just Refl <- testEquality w (knownNat @32)
-          -> checkMainWithArguments anyCfg
+           (Empty :> LLVMPointerRepr w :> PtrRepr)
+             |  isMain, shouldSupplyMainArguments
+             ,  Just Refl <- testEquality w (knownNat @32)
+             -> checkMainWithArguments anyCfg
 
-        _ -> throwCError (BadFun nm isMain)  -- TODO(lb): Suggest uc-crux-llvm?
+           _ -> throwCError (BadFun nm isMain)  -- TODO(lb): Suggest uc-crux-llvm?
+
     Nothing -> throwCError (MissingFun nm)
   where
     nm     = entryPoint llvmOpts
-    mp     = cfgMap trans
     isMain = nm == "main"
     shouldSupplyMainArguments =
       case supplyMainArguments llvmOpts of